peerstash-control/pyproject.toml
[project]
name = "peerstash"
version = "0.9.0"
description = "Peer-to-peer (p2p) backups between NAS machines owned by a group of friends"
authors = [
    { name = "BPR02", email = "dev@bpr02.com" }
]
requires-python = ">=3.13"
dependencies = [
    "commentjson>=0.9.0",
    "cron-validator>=1.0.8",
    "cryptography>=46.0.5",
    "paramiko>=4.0.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "resticpy>=1.3.0",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "pyright>=1.1.408",
    "pytest-mock>=3.15.1",
    "pytest>=9.0.2",
    "responses>=0.26.0",
    "pytest-cov>=7.0.0",
]
build = [
    "nuitka>=4.0",
    "zstandard>=0.25.0",
]

[project.scripts]
peerstash = "peerstash.cli:main"
peerstashd = "peerstash.daemon:main"
peerstash-init = "peerstash.setup:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
    packages = ["peerstash"]

[tool.pytest.ini_options]
filterwarnings = [
    # Ignore Python 3.13 deprecation warnings specifically coming from the lark library
    "ignore:module 'sre_parse' is deprecated:DeprecationWarning:lark.*",
    "ignore:module 'sre_constants' is deprecated:DeprecationWarning:lark.*"
]
